stdAcc library - really powerful for automation. This library also includes methods to get proxy children (i.e. handle children without IAccessible interfaces (at least as much as we can…)).
Methods include:
FromPoint - Get Acc object underneath a point
FromHwnd - Get Acc object for a window
FromExcel -  Get Acc object for excel
FromDesktop - Get Acc object for desktop
FromIAccessible - Get acc object from an IAccessible interface
FromMouse - Get acc object under the mouse
FromPath - A way of querying children from a direct path.
GetDescendents() - Recursively search for children and return a
collection of all descendents.
FindFirst() - Takes a query string and identifies descendents which
mach query.
FindAll() - Like FindFirst but returns a collection of stdAcc objects.
Parent - Gets the parent acc object
children - gets the children of this acc object (Collection<stdAcc>)
hwnd - gets the hwnd of this acc object (if existent)
location - gets the location of this acc object as a collection with
keys left, top, width, height and parent
HitTest - returns the stdAcc object at the coordinates specified
Selection - get the selected acc object
value - Get or set the value of an acc element.
name - get the name of an acc element.
DefaultAction -  get the default action of an acc element.
Role - get the role of  an acc element (as a text string)
state - get the state of an acc element (as a text string)
description - gets the description of an acc element
keyboardShortcut - gets the keyboard shortcut of an acc element
focus - get or set focus for an acc control.
help - gets the help text of an acc element
helptopic - gets the help topic of an acc element (which topic help is
stored in the help file)
DoDefaultAction() - performs the default action of an acc element.
SendMessage() - Send a win32 message to an acc control and return
results.

‘Protected methods for internal use
GetIAccessible - Get the IAccessible object
SetIAccessible - Set the IAccessible object