stdClipboard integration with Master (#9)
# `stdClipboard`

Regularly I've wanted to read and write data to and from the clipboard. Unfortunately with VBA there are many code snippets around for doing different functions of clipboard management, but no single point of contact to solve all clipboard management needs. Until now.

This static class offers a full set of APIs for managing the clipboard.

## API Documentation:

Property Read/Write value(ByVal iFormat As CLIPFORMAT)

The value is an advanced API which can be used to paste any format into the clipboard. Typically this API will be avoided but in some circumstances it will be your only option. Typically using this API will involve constructing a byte array and passing this in to a write value. Alternatively you can Read from any value supported by the current clipboard content, to view it.

### Property Read/Write `text` as `String`

Use this property to set the clipboard to the value of a string.

### Property Read/Write `Picture` as `stdole.IPicture`

Use this property to either inject a picture into the clipboard, or to retrieve a picture object from the clipboard. `IPicture` retrieval can be used for saving the image as a file. In combination with `XLPicture` this can be used to save pictures in the workbook to a file. This can also be used for saving charts to image files.

### Property Read/Write `files` As `Collection`

Use this property to set/get the paths to the files which are currently stored on the clipboard. Ever wanted to copy files, as if they were copied in explorer? Or maybe you wanted to "drag and drop" files onto another control. This API will help out with these kind of endeavors.

### Property WriteOnly `XLPicture` as `Excel.IPicture`

Use this property to write a `XLPicture` which is currently stored in an Excel spreadsheet, to the clipboard.

### Property ReadOnly `formats` As `Collection`

Use this property to obtain all the names of all the formats supported by the current contents of the clipboard.

### Property ReadOnly `formatIDs` As `Collection`

Use this property to obtain the CLIPFORMAT values which are supported by the current contents of the clipboard. e.g.  `formatIDs()("CF_BITMAP")`

### Function `IsFormatAvailable(ByVal iFormat As CLIPFORMAT)` As `Boolean`

Use this property to determine whether a given format is supported by the current contents of the clipboard.

### Sub `Await()`

Used to wait until the clipboard is ready.

> Note: This method is required if you are waiting for another application to copy some data into the clipboard.

### Sub `SetFilesArr(ByVal vFiles As Variant)`

Similar to setting the files property, but here you can pass an array instead.

### Sub `SendPaste(ByVal hwnd As LongPtr/Long)`

This allows you to send a paste event to a window.

> Note this is not the same as pressing Ctrl+V

### Sub `SendCopy(ByVal hwnd As LongPtr/Long)`

This allows you to send a copy event to a window.

> Note this is not the same as pressing Ctrl+C

## Helpers

### Function `GetClipFormatFromCopyPictureFormat(ByVal CopyPictureFormat As Long)` As `CLIPFORMAT`

Use this function to convert from `xlCopyPictureFormat` to `CLIPFORMAT`.

## `Protected`

Protected methods are prefixed with `prot` and typically shouldn't be used, but are exposed for if/when you need them.

## Examples

For examples see the clipboard tests.