MAJOR stdArray
# BREAKING CHANGE
We are removing callback metadata! Callback metadata used to be a means of adding an aditional argument onto a callback usually to store some additional data which could be used within the callback function. With the advent of Bind() method on stdICallable, metadata can be replaced.

Change from this convention:

```vb
arr.map(callable, someMetadata)
```

To this convention

```vb
arr.map(callable.bind(someMetadata))
```

We apologise for any inconvenience this causes. but it will pave the way for a cleaner and more easily maintainable code base!

# Other changes

* Unit testing for all array methods has been added.
* All bugs found during unit testing have been fixed.