Fix improper paramarray garbage collection issue causing crash
Sometimes `vLastArgs = params` was causing `params` to be deleted by the garbage collector... Not sure why? This was found to be fixable by calling `CopyVariant`, however it was noticed that vLastArgs as a property was no longer really required. For this reason, the property was removed and added as a parameter of `evaluate()` instead.