Addition of loops in stdlambda
2 Things are added: Do loops and for loops.
Do loops work by running an infinite loop through a jump operation until a statement is true.
For loops work by running the loop n times.
Do-Loop syntax is the same as VBAs
For-Loops are the same too, only that putting a word behind "next" will do nothing useful. Just write "next" at the end for that loop.
Both loops need all their variables defined before the loop.
If not, an error occures where the variable-name will be pushed onto stack instead of its value.
I added some new subs to work with these two loops. I did not change anything major in the operation-creation.
Only exception to that is how for loops are interpreted, as they requiere more inline creation of tokens. (for example it needs i = i + 1 to increment its index, otherwise it would run infinetly.
Mind the line 1082.
This is a proof of concept, feedback would be nice.